requirements.txt
psycopg2-binary==2.9.10
Flask
Flask-SQLAlchemy
flask-babel==4.0.0
pandas
openpyxl
pillow==11.3.0
